Add default values for lft, rght, tree_id and level
Because of the way the model is imported, it doesn't fill in the values for lft, rght, and tree_id (which cannot be None) like the model manager does if it's used in a shell or outside of a migration. Setting the default values does the same as initializing one node as the root in a shell.